- Included dynamic-import header to tomcat-jdbc-8.0.30.jar - Added HikariCP sample implementation - Refined the package structure.
