+++ modules/carbon-datasources/org.wso2.carbon.datasource.core/pom.xml
@@ -11,7 +11,7 @@
 
     <groupId>org.wso2.carbon</groupId>
     <artifactId>org.wso2.carbon.datasource.core</artifactId>
-    <version>1.0-SNAPSHOT</version>
+    <version>1.0.0</version>
     <packaging>bundle</packaging>
 
     <dependencies>
@@ -45,6 +45,11 @@
             <artifactId>biz.aQute.bndlib</artifactId>
             <version>3.0.0</version>
         </dependency>
+        <dependency>
+            <groupId>com.zaxxer</groupId>
+            <artifactId>HikariCP</artifactId>
+            <version>2.4.1</version>
+        </dependency>
     </dependencies>
 
     <build>
@@ -72,9 +77,7 @@
                         <Bundle-Activator>org.wso2.carbon.datasource.core.DataSourceBundleActivator</Bundle-Activator>
                         <Bundle-SymbolicName>${project.artifactId}</Bundle-SymbolicName>
                         <Bundle-Name>${project.artifactId}</Bundle-Name>
-                        <Private-Package>org.wso2.carbon.datasource.core.internal</Private-Package>
                         <Export-Package>
-                            !org.wso2.carbon.datasource.internal,
                             org.wso2.carbon.datasource.core.*,org.wso2.carbon.datasource.rdbms.*,
                             org.wso2.carbon.datasource.common.*,org.wso2.carbon.datasource.utils.*
                         </Export-Package>
