+++ modules/carbon-datasources/org.wso2.carbon.datasource.core/src/main/java/org/wso2/carbon/datasource/Runner.java
@@ -15,17 +15,71 @@
  */
 package org.wso2.carbon.datasource;
 
+import org.apache.tomcat.jdbc.pool.DataSource;
 import org.wso2.carbon.datasource.common.DataSourceException;
 import org.wso2.carbon.datasource.core.DataSourceManager;
 
 import java.nio.file.Paths;
+import java.sql.Connection;
+import java.sql.ResultSet;
+import java.sql.SQLException;
+import java.sql.Statement;
 import java.util.Hashtable;
 import javax.naming.Context;
 import javax.naming.InitialContext;
 import javax.naming.NamingException;
-import javax.sql.DataSource;
 
 public class Runner {
+    // Random sample to test jdbc pool
+//    public static void main(String[] args) {
+//        PoolProperties p = new PoolProperties();
+//        p.setUrl("jdbc:mysql://localhost:3306/test");
+//        p.setDriverClassName("com.mysql.jdbc.Driver");
+//        p.setUsername("root");
+//        p.setPassword("root");
+//        p.setJmxEnabled(true);
+//        p.setTestWhileIdle(false);
+//        p.setTestOnBorrow(true);
+//        p.setValidationQuery("SELECT 1");
+//        p.setTestOnReturn(false);
+//        p.setValidationInterval(30000);
+//        p.setTimeBetweenEvictionRunsMillis(30000);
+//        p.setMaxActive(100);
+//        p.setInitialSize(10);
+//        p.setMaxWait(10000);
+//        p.setRemoveAbandonedTimeout(60);
+//        p.setMinEvictableIdleTimeMillis(30000);
+//        p.setMinIdle(10);
+//        p.setLogAbandoned(true);
+//        p.setRemoveAbandoned(true);
+//        p.setJdbcInterceptors("org.apache.tomcat.jdbc.pool.interceptor.ConnectionState;"+
+//                "org.apache.tomcat.jdbc.pool.interceptor.StatementFinalizer");
+//        DataSource datasource = new DataSource();
+//        datasource.setPoolProperties(p);
+//
+//        Connection con = null;
+//        try {
+//            con = datasource.getConnection();
+//            Statement st = con.createStatement();
+//            ResultSet rs = st.executeQuery("select * from pet");
+//            int cnt = 1;
+//            while (rs.next()) {
+//                System.out.println((cnt++)+". Host:" +rs.getString("name")+
+//                        " User:"+rs.getString("owner")+" Password:"+rs.getString("birth"));
+////                System.out.println((cnt++)+". Host:" +rs.getString("Host")+
+////                        " User:"+rs.getString("User")+" Password:"+rs.getString("Password"));
+//            }
+//            rs.close();
+//            st.close();
+//        } catch (SQLException e) {
+//            e.printStackTrace();
+//        } finally {
+//            if (con!=null) try {con.close();}catch (Exception ignore) {}
+//        }
+//    }
+
+
+    //Sample to bind and fetch datasource and query for data
     public static void main(String[] args) {
         DataSourceManager manager = DataSourceManager.getInstance();
 
@@ -43,11 +97,34 @@ public class Runner {
             map.put(Context.PROVIDER_URL, "file:///home/dinushab/Carbon/POC/data-sources/carbon-kernel/modules/carbon-datasources/org.wso2.carbon.datasource.core/src/test");
             InitialContext context = new InitialContext(map);
             DataSource dataSource = (DataSource)context.lookup("jdbc/WSO2CarbonDB/test");
-            System.out.println(dataSource.toString());
+
+
+            Connection con = null;
+            try {
+                con = dataSource.getConnection();
+                Statement st = con.createStatement();
+                ResultSet rs = st.executeQuery("select * from pet");
+                int cnt = 1;
+                while (rs.next()) {
+                    System.out.println((cnt++)+". First name: " +rs.getString("name")+
+                            ", Last name: "+rs.getString("owner")+", DOB:"+rs.getString("birth"));
+//                System.out.println((cnt++)+". Host:" +rs.getString("Host")+
+//                        " User:"+rs.getString("User")+" Password:"+rs.getString("Password"));
+                }
+                rs.close();
+                st.close();
+            } catch (SQLException e) {
+                e.printStackTrace();
+            } finally {
+                if (con!=null) try {con.close();}catch (Exception ignore) {}
+            }
+
         } catch (DataSourceException e) {
             e.printStackTrace();
         } catch (NamingException e) {
             e.printStackTrace();
         }
     }
+
+
 }
