+++ modules/carbon-datasources/org.wso2.carbon.datasource.core/src/main/java/org/wso2/carbon/datasource/rdbms/tomcat/ConnectionRollbackOnReturnInterceptor.java
@@ -13,7 +13,7 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-package org.wso2.carbon.datasource.rdbms;
+package org.wso2.carbon.datasource.rdbms.tomcat;
 
 import org.apache.tomcat.jdbc.pool.ConnectionPool;
 import org.apache.tomcat.jdbc.pool.JdbcInterceptor;
