+++ modules/carbon-datasources/org.wso2.carbon.datasource.core/src/main/java/org/wso2/carbon/datasource/rdbms/tomcat/RDBMSDataSourceReader.java
@@ -1,0 +1,105 @@
+/*
+ *  Copyright (c) 2016, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *  http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
+package org.wso2.carbon.datasource.rdbms.tomcat;
+
+import org.wso2.carbon.datasource.common.DataSourceException;
+import org.wso2.carbon.datasource.common.spi.DataSourceReader;
+import org.wso2.carbon.datasource.rdbms.RDBMSConfiguration;
+import org.wso2.carbon.datasource.rdbms.RDBMSDataSourceConstants;
+import org.wso2.carbon.datasource.rdbms.utils.RDBMSDataSourceUtils;
+
+import java.sql.Connection;
+import java.sql.DriverManager;
+import java.sql.PreparedStatement;
+import java.sql.SQLException;
+import javax.sql.DataSource;
+
+/**
+ * This class represents the RDBMS based data source reader implementation.
+ */
+public class RDBMSDataSourceReader implements DataSourceReader {
+
+    @Override
+    public String getType() {
+        return RDBMSDataSourceConstants.RDBMS_DATASOURCE_TYPE;
+    }
+
+
+    @Override
+    public Object createDataSource(String xmlConfiguration, boolean isDataSourceFactoryReference)
+            throws DataSourceException {
+        RDBMSDataSource dataSource = new RDBMSDataSource(RDBMSDataSourceUtils.loadConfig(xmlConfiguration));
+        if (isDataSourceFactoryReference) {
+            return dataSource.getDataSourceFactoryReference();
+        } else {
+            return dataSource.getDataSource();
+        }
+    }
+
+    @Override
+    public boolean testDataSourceConnection(String xmlConfiguration) throws DataSourceException {
+        RDBMSConfiguration rdbmsConfiguration = RDBMSDataSourceUtils.loadConfig(xmlConfiguration);
+        DataSource dataSource = new RDBMSDataSource(rdbmsConfiguration).getDataSource();
+
+        Connection connection = null;
+        try {
+            Class.forName(rdbmsConfiguration.getDriverClassName());
+            if (rdbmsConfiguration.getUsername() != null) {
+                DriverManager.getConnection(rdbmsConfiguration.getUrl(), rdbmsConfiguration.getUsername(),
+                        rdbmsConfiguration.getPassword());
+            } else {
+                DriverManager.getConnection(rdbmsConfiguration.getUrl());
+            }
+        } catch (ClassNotFoundException e) {
+            throw new DataSourceException("Error loading Driver class:" + e.getMessage(), e);
+        } catch (SQLException e) {
+            if (e.getSQLState().equals("08001")) {
+                throw new DataSourceException("The data source URL is not accepted by any of the loaded drivers. "
+                        + e.getMessage(), e);
+            } else if (e.getSQLState().equals("28000")) {
+                throw new DataSourceException("The user is not associated with a trusted SQL Server connection."
+                        + e.getMessage(), e);
+            } else {
+                throw new DataSourceException("Error establishing data source connection: " + e.getMessage(), e);
+            }
+        }
+        try {
+            connection = dataSource.getConnection();
+        } catch (SQLException e) {
+            throw new DataSourceException("Error establishing data source connection: " + e.getMessage(), e);
+        }
+        if (connection != null) {
+            String validationQuery = rdbmsConfiguration.getValidationQuery();
+            if (validationQuery != null && !"".equals(validationQuery)) {
+                PreparedStatement ps = null;
+                try {
+                    ps = connection.prepareStatement(validationQuery.trim());
+                    ps.execute();
+                    ps.close();
+                } catch (SQLException e) {
+                    throw new DataSourceException("Error during executing validation query: " + e.getMessage(), e);
+                }
+            }
+            try {
+                connection.close();
+            } catch (SQLException ignored) {
+
+            }
+        }
+        return true;
+    }
+
+}
